move processors into folder impl
