+++ pkg/plugin/processor/builtin/impl/field/rename_examples_test.go
